backend/requirements.txt
pandas
requests
beautifulsoup4
rapidfuzz
pyyaml
openpyxl
streamlit
plotly
